code source/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
scikit-learn
tensorflow
keras-tcn
matplotlib
seaborn


#pip install transformers accelerate
